Move MMConnect tests to Appium smoke CI under SmokeMMConnect.
Migrate the suite from performance BrowserStack to PR-blocking Appium smoke, wire appium-mmconnect-android-smoke, and remove mm-connect from the performance workflow.
