+++ .github/workflows/performance-test-runner.yml
@@ -10,7 +10,7 @@ on:
       build_type:
         required: true
         type: string
-        description: 'Build type (imported-wallet/onboarding/mm-connect)'
+        description: 'Build type (imported-wallet/onboarding)'
       device_matrix:
         required: true
         type: string
@@ -31,10 +31,6 @@ on:
         required: true
         type: string
         description: 'Unified BrowserStack build name for all sessions'
-      browserstack_playground_url:
-        required: false
-        type: string
-        description: 'BrowserStack URL for the RN playground APK (mm-connect tests only)'
       sentry_target:
         required: false
         type: string
@@ -136,41 +132,6 @@ jobs:
           access-key: ${{ secrets.BROWSERSTACK_ACCESS_KEY }}
           project-name: ${{ github.repository }}
 
-      - name: Compute BrowserStack Local Identifier
-        id: bs-local-id
-        if: inputs.build_type == 'mm-connect'
-        run: |
-          # Strip any character that is not alphanumeric or a hyphen (covers spaces, /, \, @, etc.)
-          BUILD_TYPE_SAFE=$(printf '%s' "${{ inputs.build_type }}" | tr -c 'a-zA-Z0-9-' '_')
-          DEVICE_SAFE=$(printf '%s' "${{ matrix.device.name }}" | tr -c 'a-zA-Z0-9-' '_')
-          if [ -z "$BUILD_TYPE_SAFE" ]; then
-            echo "❌ Error: inputs.build_type is empty — cannot build a valid BrowserStack Local identifier"
-            exit 1
-          fi
-          if [ -z "$DEVICE_SAFE" ]; then
-            echo "❌ Error: matrix.device.name is empty — cannot build a valid BrowserStack Local identifier"
-            exit 1
-          fi
-          echo "value=${{ github.run_id }}-${BUILD_TYPE_SAFE}-${DEVICE_SAFE}" >> "$GITHUB_OUTPUT"
-
-      - name: Setup BrowserStack Local
-        if: inputs.build_type == 'mm-connect'
-        uses: browserstack/github-actions/setup-local@4478e16186f38e5be07721931642e65a028713c3
-        with:
-          local-testing: start
-          local-identifier: ${{ steps.bs-local-id.outputs.value }}
-          # Use same args for all build types. Do not use --include-hosts for mm-connect:
-          # bs-local.com requests must be forwarded to localhost; --include-hosts can block them.
-          local-args: '--force-local --verbose'
-
-      - name: Wait for BrowserStack Local
-        if: inputs.build_type == 'mm-connect'
-        run: |
-          echo "Waiting for BrowserStack Local to be ready..."
-          # mm-connect needs the tunnel ready so the device can reach bs-local.com:8090
-          sleep 15
-          echo "BrowserStack Local should be ready now"
-
       - name: Set Test Environment
         run: |
           echo "Setting ${{ inputs.build_type }} test environment for device: ${{ matrix.device.name }} (${{ matrix.device.category }} Class)"
@@ -220,9 +181,6 @@ jobs:
             echo "BROWSERSTACK_DEVICE=${{ matrix.device.name }}"
             echo "BROWSERSTACK_OS_VERSION=${{ matrix.device.os_version }}"
             echo "BROWSERSTACK_${{ inputs.platform == 'android' && 'ANDROID' || 'IOS' }}_${{ inputs.build_type == 'onboarding' && 'CLEAN_' || '' }}APP_URL=${{ inputs.browserstack_app_url }}"
-            if [ -n "${{ inputs.browserstack_playground_url }}" ]; then
-              echo "BROWSERSTACK_RN_PLAYGROUND_URL=${{ inputs.browserstack_playground_url }}"
-            fi
             echo "TEST_PLATFORM=${{ inputs.platform }}"
             echo "QA_APP_VERSION=${{ inputs.app_version }}"
             echo "BROWSERSTACK_BUILD_NAME=${{ inputs.browserstack_build_name }}"
@@ -245,8 +203,6 @@ jobs:
       - name: Run Tests
         continue-on-error: true
         env:
-          BROWSERSTACK_LOCAL: ${{ inputs.build_type == 'mm-connect' && 'true' || 'false' }}
-          BROWSERSTACK_LOCAL_IDENTIFIER: ${{ inputs.build_type == 'mm-connect' && steps.bs-local-id.outputs.value || '' }}
           BROWSERSTACK_BUILD_NAME: ${{ inputs.browserstack_build_name }}
           GREP_TAGS: ${{ inputs.grep_tags }}
         working-directory: '.'
@@ -266,8 +222,6 @@ jobs:
           # Select base command based on build_type
           if [ "${{ inputs.build_type }}" = "onboarding" ]; then
             BASE_CMD="yarn run-playwright:${{ inputs.platform }}-onboarding-bs"
-          elif [ "${{ inputs.build_type }}" = "mm-connect" ]; then
-            BASE_CMD="yarn run-playwright:mm-connect-${{ inputs.platform }}-bs"
           else
             BASE_CMD="yarn run-playwright:${{ inputs.platform }}-bs"
           fi
