+++ .github/workflows/run-appium-smoke-tests-android.yml
@@ -78,3 +78,27 @@ jobs:
       metamask_environment: ${{ inputs.metamask_environment }}
       runner_provider: ${{ inputs.runner_provider }}
     secrets: inherit
+
+  appium-mmconnect-android-smoke:
+    if: >-
+      ${{
+        !cancelled() &&
+        (contains(fromJson(inputs.selected_tags), 'ALL') ||
+         contains(fromJson(inputs.selected_tags), 'SmokeMMConnect'))
+      }}
+    strategy:
+      matrix:
+        split: [1]
+      fail-fast: false
+    uses: ./.github/workflows/run-appium-e2e-workflow.yml
+    with:
+      test-suite-name: appium-mmconnect-android-smoke-${{ matrix.split }}
+      platform: android
+      test_suite_tag: SmokeMMConnect
+      split_number: ${{ matrix.split }}
+      total_splits: 1
+      test-timeout-minutes: 35
+      build_type: ${{ inputs.build_type }}
+      metamask_environment: ${{ inputs.metamask_environment }}
+      runner_provider: ${{ inputs.runner_provider }}
+    secrets: inherit
