+++ .github/workflows/run-performance-e2e.yml
@@ -163,7 +163,6 @@ jobs:
     needs: [determine-branch-name]
     outputs:
       android_matrix: ${{ steps.read-matrix.outputs.android_matrix }}
-      android_mm_connect_matrix: ${{ steps.read-matrix.outputs.android_mm_connect_matrix }}
       ios_matrix: ${{ steps.read-matrix.outputs.ios_matrix }}
     steps:
       - name: Checkout code
@@ -182,17 +181,14 @@ jobs:
           # Schedule and workflow_dispatch use the full device matrix.
           if [ -n "$PR_NUMBER" ]; then
             ANDROID_MATRIX=$(jq '[.android_devices[] | select(.category == "low")]' "$FILE")
-            ANDROID_MM_CONNECT_MATRIX="$ANDROID_MATRIX"
             IOS_MATRIX='[]'
             echo "PR mode: Android low category devices only (PR: $PR_NUMBER, branch: $BRANCH_NAME)"
           elif [ "${{ github.event_name }}" = "push" ]; then
             ANDROID_MATRIX=$(jq '[.android_devices[] | select(.category == "low")]' "$FILE")
-            ANDROID_MM_CONNECT_MATRIX=$(jq '[.android_devices[] | select(.name | contains("Samsung"))]' "$FILE")
             IOS_MATRIX=$(jq '[.ios_devices[] | select(.category == "low")]' "$FILE")
             echo "Push mode: low category devices only (event: ${{ github.event_name }}, branch: $BRANCH_NAME)"
           else
             ANDROID_MATRIX=$(jq ".android_devices" "$FILE")
-            ANDROID_MM_CONNECT_MATRIX=$(jq '[.android_devices[] | select(.name | contains("Samsung"))]' "$FILE")
             IOS_MATRIX=$(jq ".ios_devices" "$FILE")
             echo "Full mode: All devices (event: ${{ github.event_name }}, branch: $BRANCH_NAME)"
           fi
@@ -201,16 +197,12 @@ jobs:
             echo "android_matrix<<EOF"
             echo "$ANDROID_MATRIX"
             echo "EOF"
-            echo "android_mm_connect_matrix<<EOF"
-            echo "$ANDROID_MM_CONNECT_MATRIX"
-            echo "EOF"
             echo "ios_matrix<<EOF"
             echo "$IOS_MATRIX"
             echo "EOF"
           } >> "$GITHUB_OUTPUT"
 
           echo "Selected: $(echo "$ANDROID_MATRIX" | jq length) Android, $(echo "$IOS_MATRIX" | jq length) iOS"
-          echo "Selected for Android MM-Connect: $(echo "$ANDROID_MM_CONNECT_MATRIX" | jq length)"
 
   set-build-names:
     name: Set Unified BrowserStack Build Names
@@ -421,69 +413,6 @@ jobs:
       grep_tags: ${{ needs.compute-test-selection.outputs.grep_pattern }}
     secrets: inherit
 
-  fetch-rn-playground-apk-upload-to-browserstack:
-    name: Fetch RN Playground APK and Upload to BrowserStack
-    runs-on: ubuntu-latest
-    needs: [wait-for-onboarding-completion]
-    if: always() && !cancelled()
-    outputs:
-      browserstack-playground-url: ${{ steps.upload-playground.outputs.browserstack-url }}
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v4
-
-      - name: Fetch playground APK from GitHub Releases
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          RN_PLAYGROUND_APK_VERSION: ${{ vars.RN_PLAYGROUND_APK_VERSION || '' }}
-        run: ./tests/scripts/fetch-rn-playground-apk.sh --output ./tmp/rn-playground.apk
-
-      - name: Upload playground APK to BrowserStack
-        id: upload-playground
-        run: |
-          echo "Uploading RN Playground APK to BrowserStack..."
-          RESPONSE=$(curl -u "$BROWSERSTACK_USERNAME:$BROWSERSTACK_ACCESS_KEY" \
-            -X POST "https://api-cloud.browserstack.com/app-automate/upload" \
-            -F "file=@./tmp/rn-playground.apk" \
-            -F "custom_id=RN-Playground-${{ github.run_id }}")
-
-          APP_URL=$(echo "$RESPONSE" | jq -r '.app_url')
-          if [ -z "$APP_URL" ] || [ "$APP_URL" = "null" ]; then
-            echo "Error: Failed to upload playground APK to BrowserStack"
-            echo "Response: $RESPONSE"
-            exit 1
-          fi
-
-          echo "browserstack-url=$APP_URL" >> "$GITHUB_OUTPUT"
-          echo "Playground APK uploaded: $APP_URL"
-
-  run-android-mm-connect-tests:
-    name: Run Android MM-Connect Tests
-    uses: ./.github/workflows/performance-test-runner.yml
-    needs:
-      [
-        read-device-matrix,
-        trigger-android-dual-versions,
-        fetch-rn-playground-apk-upload-to-browserstack,
-        set-build-names,
-        determine-branch-name,
-        compute-test-selection,
-      ]
-    if: always() && !cancelled() && needs.compute-test-selection.outputs.run_tests == 'true' && (needs.trigger-android-dual-versions.result == 'skipped' || needs.trigger-android-dual-versions.result == 'success') && (inputs.browserstack_app_url_android_imported_wallet != '' || needs.trigger-android-dual-versions.outputs.with-srp-browserstack-url != '')
-    with:
-      platform: android
-      build_type: mm-connect
-      sentry_target: ${{ inputs.sentry_target || 'test' }}
-      build_variant: ${{ inputs.build_variant || 'e2e' }}
-      device_matrix: ${{ needs.read-device-matrix.outputs.android_mm_connect_matrix }}
-      browserstack_app_url: ${{ needs.trigger-android-dual-versions.outputs.with-srp-browserstack-url || inputs.browserstack_app_url_android_imported_wallet }}
-      app_version: ${{ needs.trigger-android-dual-versions.outputs.with-srp-version || 'Manual-Input' }}
-      branch_name: ${{ needs.determine-branch-name.outputs.branch_name }}
-      browserstack_build_name: ${{ needs.set-build-names.outputs.android_build_name }}
-      browserstack_playground_url: ${{ needs.fetch-rn-playground-apk-upload-to-browserstack.outputs.browserstack-playground-url }}
-      grep_tags: ${{ needs.compute-test-selection.outputs.grep_pattern }}
-    secrets: inherit
-
   aggregate-results:
     name: Aggregate All Test Results
     runs-on: ubuntu-latest
@@ -491,7 +420,6 @@ jobs:
       [
         run-android-imported-wallet-tests,
         run-android-onboarding-tests,
-        run-android-mm-connect-tests,
         run-ios-imported-wallet-tests,
         run-ios-onboarding-tests,
         wait-for-onboarding-completion,
@@ -588,7 +516,6 @@ jobs:
       [
         run-android-imported-wallet-tests,
         run-android-onboarding-tests,
-        run-android-mm-connect-tests,
         run-ios-imported-wallet-tests,
         run-ios-onboarding-tests,
         aggregate-results,
