+++ package.json
@@ -120,9 +120,6 @@
     "run-playwright:ios-bs": "yarn playwright test --project browserstack-ios --config tests/playwright.config.ts",
     "run-playwright:android": "yarn playwright test --project android --config tests/playwright.config.ts",
     "run-playwright:ios": "yarn playwright test --project ios --config tests/playwright.config.ts",
-    "run-playwright:mm-connect-android-local": "yarn playwright test --project mm-connect-android-local --config tests/playwright.config.ts",
-    "run-playwright:mm-connect-android-bs": "yarn playwright test --project mm-connect-android-browserstack --config tests/playwright.config.ts",
-    "run-playwright:mm-connect-android-bs-local": "yarn playwright test --project mm-connect-android-browserstack --config tests/playwright.config.ts",
     "run-system-tests:android-login": "yarn playwright test --project system-android-login --config tests/playwright.system.config.ts",
     "run-system-tests:android-onboarding": "yarn playwright test --project system-android-onboarding --config tests/playwright.system.config.ts",
     "run-system-tests:ios-login": "yarn playwright test --project system-ios-login --config tests/playwright.system.config.ts",
@@ -133,6 +130,7 @@
     "run-system-tests:ios-login-sim": "yarn playwright test --project system-ios-login-sim --config tests/playwright.system-emulator.config.ts",
     "run-system-tests:ios-onboarding-sim": "yarn playwright test --project system-ios-onboarding-sim --config tests/playwright.system-emulator.config.ts",
     "appium-smoke:android": "yarn playwright test --config tests/playwright.smoke-appium.config.ts --project android-smoke",
+    "appium-smoke:mmconnect:android": "yarn playwright test --config tests/playwright.smoke-appium.config.ts --project android-smoke --grep SmokeMMConnect",
     "appium-smoke:ios": "yarn playwright test --config tests/playwright.smoke-appium.config.ts --project ios-smoke",
     "capture-visual-baselines:android": "CAPTURE_BASELINES=true AI_VISUAL_TESTING_ENABLED=true yarn playwright test --project system-android-login-emu --config tests/playwright.system-emulator.config.ts",
     "capture-visual-baselines:ios": "CAPTURE_BASELINES=true AI_VISUAL_TESTING_ENABLED=true yarn playwright test --project system-ios-login-sim --config tests/playwright.system-emulator.config.ts",
