--- tests/performance/mm-connect/connection-evm-sign.spec.ts
@@ -1,232 +1,0 @@
-import { test } from '../../framework/fixtures/playwright';
-import { Performance } from '../../tags.performance.js';
-
-import { loginToAppPlaywright } from '../../flows/wallet.flow';
-import BrowserPlaygroundDapp from '../../page-objects/MMConnect/BrowserPlaygroundDapp';
-import AndroidScreenHelpers from '../../page-objects/MMConnect/AndroidScreenHelpers';
-import DappConnectionModal from '../../page-objects/MMConnect/DappConnectionModal';
-import SignModal from '../../page-objects/MMConnect/SignModal';
-import SwitchChainModal from '../../page-objects/MMConnect/SwitchChainModal';
-import PlaywrightContextHelpers from '../../framework/PlaywrightContextHelpers';
-import { DappServer, DappVariants, TestDapps, sleep } from '../../framework';
-import {
-  getDappUrlForBrowser,
-  setupAdbReverse,
-  cleanupAdbReverse,
-  waitForDappServerReady,
-  unlockIfLockScreenVisible,
-  ensureAccountGroupsFinishedLoading,
-} from './utils';
-import {
-  launchMobileBrowser,
-  navigateToDapp,
-  switchToMobileBrowser,
-} from '../../flows/native-browser.flow';
-
-const DAPP_PORT = 8090;
-
-// NOTE: This test requires the testing SRP to be used
-const ACCOUNT_1_ADDRESS = '0x19a7Ad8256ab119655f1D758348501d598fC1C94';
-
-const playgroundServer = new DappServer({
-  dappCounter: 0,
-  rootDirectory: TestDapps[DappVariants.BROWSER_PLAYGROUND].dappPath,
-  dappVariant: DappVariants.BROWSER_PLAYGROUND,
-});
-
-test.describe(Performance, () => {
-  test.beforeAll(async () => {
-    playgroundServer.setServerPort(DAPP_PORT);
-    await playgroundServer.start();
-    await waitForDappServerReady(DAPP_PORT);
-    setupAdbReverse(DAPP_PORT);
-  });
-
-  test.afterAll(async () => {
-    cleanupAdbReverse(DAPP_PORT);
-    await playgroundServer.stop();
-  });
-
-  // Test steps (in order):
-  //
-  // 1. LOGIN AND NAVIGATE TO DAPP
-  //    - Login to app, ensure account groups finished loading
-  //    - Launch mobile browser and navigate to the playground dapp
-  //
-  // 2. CONNECT VIA LEGACY EVM (WITH ACCOUNT 3)
-  //    - Tap Connect (Legacy)
-  //    - In MetaMask: tap Edit Accounts, add Account 3, tap Update, tap Connect (cooldown 2s)
-  //    - Assert: connected true, chainId '0x1', active account is Account 1
-  //      (0x19a7Ad8256ab119655f1D758348501d598fC1C94)
-  //
-  // 3. SIGN MESSAGE ON ETHEREUM — CONFIRM
-  //    - Tap personal sign
-  //    - In MetaMask: tap Confirm (cooldown 2s)
-  //    - Assert: response value matches Account 1 signature hash
-  //      (0x361c13288b4ab02d50974efddf9e4e7ca651b81c298b614be908c4754abb1dd8
-  //       328224645a1a8d0fab561c4b855c7bdcebea15db5ae8d1778a1ea791dbd05c2a1b)
-  //
-  // 4. SEND TRANSACTION ON ETHEREUM — CANCEL
-  //    - Tap send transaction
-  //    - In MetaMask: assert network text 'Ethereum', tap Cancel (cooldown 2s)
-  //    - Assert: response value 'denied'
-  //
-  // 5. SWITCH TO POLYGON AND SEND TRANSACTION — CANCEL
-  //    - Assert response 'denied', tap switch to Polygon
-  //    - In MetaMask: assert network text 'Polygon', tap Confirm chain switch (cooldown 2s)
-  //    - Assert: chainId '0x89'
-  //    - Tap send transaction
-  //    - In MetaMask: assert network text 'Polygon', tap Cancel (cooldown 2s)
-  //
-  // 6. SWITCH TO MAINNET AND SEND TRANSACTION — CANCEL
-  //    - Tap switch to Mainnet, assert chainId '0x1'
-  //    - Tap send transaction
-  //    - In MetaMask: assert network text 'Ethereum', tap Cancel (cooldown 2s)
-  //
-  // 7. CLEANUP
-  //    - Tap disconnect to reset dapp state
-  // This test is currently being skipped as it is flaky - https://consensyssoftware.atlassian.net/browse/WAPI-1511
-  test.skip('@metamask/connect-evm - Sign and transaction cancel flows', async ({
-    currentDeviceDetails,
-    driver,
-  }) => {
-    const platform = currentDeviceDetails.platform;
-    const useBrowserStackLocal =
-      process.env.BROWSERSTACK_LOCAL?.toLowerCase() === 'true';
-    const DAPP_URL = useBrowserStackLocal
-      ? `http://bs-local.com:${DAPP_PORT}`
-      : getDappUrlForBrowser(platform);
-
-    await PlaywrightContextHelpers.withNativeAction(async () => {
-      await loginToAppPlaywright();
-      await ensureAccountGroupsFinishedLoading(currentDeviceDetails);
-      await launchMobileBrowser();
-      await navigateToDapp(DAPP_URL);
-    });
-    await sleep(5000);
-
-    await PlaywrightContextHelpers.withWebAction(async () => {
-      await BrowserPlaygroundDapp.tapConnectLegacy();
-    }, DAPP_URL);
-
-    await PlaywrightContextHelpers.withNativeAction(async () => {
-      await AndroidScreenHelpers.tapOpenDeeplinkWithMetaMask();
-      await unlockIfLockScreenVisible();
-      await DappConnectionModal.tapEditAccountsButton();
-      await DappConnectionModal.tapAccountButton('Account 3');
-      await DappConnectionModal.tapUpdateAccountsButton();
-      await DappConnectionModal.tapConnectButton({
-        shouldCooldown: true,
-        timeToCooldown: 2000,
-      });
-    });
-
-    await sleep(1000);
-    await switchToMobileBrowser();
-    await sleep(1000);
-
-    await PlaywrightContextHelpers.withWebAction(async () => {
-      await BrowserPlaygroundDapp.assertConnected(true);
-      await BrowserPlaygroundDapp.assertChainIdValue('0x1');
-      await BrowserPlaygroundDapp.assertActiveAccount(ACCOUNT_1_ADDRESS);
-      await BrowserPlaygroundDapp.tapPersonalSign();
-    }, DAPP_URL);
-
-    await PlaywrightContextHelpers.withNativeAction(async () => {
-      await AndroidScreenHelpers.tapOpenDeeplinkWithMetaMask();
-      await SignModal.tapConfirmButton({
-        shouldCooldown: true,
-        timeToCooldown: 2000,
-      });
-    });
-
-    await sleep(1000);
-    await switchToMobileBrowser();
-    await sleep(1000);
-
-    await PlaywrightContextHelpers.withWebAction(async () => {
-      await BrowserPlaygroundDapp.assertResponseValue(
-        // Account 1 signed the message
-        '0x361c13288b4ab02d50974efddf9e4e7ca651b81c298b614be908c4754abb1dd8328224645a1a8d0fab561c4b855c7bdcebea15db5ae8d1778a1ea791dbd05c2a1b',
-      );
-      await BrowserPlaygroundDapp.tapSendTransaction();
-    }, DAPP_URL);
-
-    await PlaywrightContextHelpers.withNativeAction(async () => {
-      await AndroidScreenHelpers.tapOpenDeeplinkWithMetaMask();
-      await SignModal.assertNetworkText('Ethereum');
-      await SignModal.tapCancelButton({
-        shouldCooldown: true,
-        timeToCooldown: 2000,
-      });
-    });
-
-    await sleep(1000);
-    await switchToMobileBrowser();
-    await sleep(1000);
-
-    await PlaywrightContextHelpers.withWebAction(async () => {
-      // Note: Error message may differ slightly in browser playground
-      await BrowserPlaygroundDapp.assertResponseValue('denied');
-      await BrowserPlaygroundDapp.tapSwitchToPolygon();
-    }, DAPP_URL);
-
-    await PlaywrightContextHelpers.withNativeAction(async () => {
-      await AndroidScreenHelpers.tapOpenDeeplinkWithMetaMask();
-      await SwitchChainModal.assertNetworkText('Polygon');
-      await SwitchChainModal.tapConnectButton({
-        shouldCooldown: true,
-        timeToCooldown: 2000,
-      });
-    });
-
-    await sleep(1000);
-    await switchToMobileBrowser();
-    await sleep(1000);
-
-    await PlaywrightContextHelpers.withWebAction(async () => {
-      await BrowserPlaygroundDapp.assertChainIdValue('0x89');
-      await BrowserPlaygroundDapp.tapSendTransaction();
-    }, DAPP_URL);
-
-    await PlaywrightContextHelpers.withNativeAction(async () => {
-      await AndroidScreenHelpers.tapOpenDeeplinkWithMetaMask();
-      await SignModal.assertNetworkText('Polygon');
-      await SignModal.tapCancelButton({
-        shouldCooldown: true,
-        timeToCooldown: 2000,
-      });
-    });
-
-    await sleep(1000);
-    await switchToMobileBrowser();
-    await sleep(1000);
-
-    await PlaywrightContextHelpers.withWebAction(async () => {
-      await BrowserPlaygroundDapp.tapSwitchToMainnet();
-      await BrowserPlaygroundDapp.assertChainIdValue('0x1');
-      await BrowserPlaygroundDapp.tapSendTransaction();
-    }, DAPP_URL);
-
-    await PlaywrightContextHelpers.withNativeAction(async () => {
-      await AndroidScreenHelpers.tapOpenDeeplinkWithMetaMask();
-      await SignModal.assertNetworkText('Ethereum');
-      await SignModal.tapCancelButton({
-        shouldCooldown: true,
-        timeToCooldown: 2000,
-      });
-    });
-
-    await sleep(1000);
-    await switchToMobileBrowser();
-    await sleep(1000);
-
-    //
-    // Reset dapp state
-    //
-
-    await PlaywrightContextHelpers.withWebAction(async () => {
-      await BrowserPlaygroundDapp.tapDisconnect();
-    }, DAPP_URL);
-  });
-}); // end describe
